% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_situations.R
\name{add_situation_col}
\alias{add_situation_col}
\title{Add situation as column}
\usage{
add_situation_col(dot_args, obs, obs_sd = NULL)
}
\arguments{
\item{dot_args}{A list (each element= version) of a list
(each element= situation) of simulations \code{data.frame}s}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s
(named by situation)}

\item{obs_sd}{A list (each element= situation) of \code{data.frame}s for the
standard deviation of the observations (named by situation)}
}
\value{
A list of three elements: the updated dot_args, obs and obs_sd.
}
\description{
Adds the situation as a column of the \code{data.frame}, i.e. the name of the list
element also becomes a column of the element.
}
\keyword{internal}
