% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{aesthetics}
\alias{aesthetics}
\title{Manages the aesthetics of the graphics}
\usage{
aesthetics(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  overlap = NULL,
  several_sit = FALSE,
  shape_sit = c("none", "txt", "symbol", "group"),
  one_version = TRUE,
  first_sim = TRUE
)
}
\arguments{
\item{sim}{A simulation list of data.frames}

\item{obs}{An observation list of data.frames}

\item{type}{The type of plot required, either "dynamic" or "scatter"}

\item{overlap}{A list of lists containing the variables to represent on the same graph
when \code{type = "dynamic"}}

\item{obs_sd}{A list of observation standard deviation data.frames}

\item{successive}{A list of lists containing the situations to be represented as a contiguous sequence
when \code{type = "dynamic"} (implies that the situations are correctly ordered)}
}
\value{
An aesthetics expression which will allow to manage the aesthetics of the graphics
}
\description{
Manages the shape, color and line type of the graphics according to their content
}
\keyword{internal}
