% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_bind_split.R
\name{bind_rows_sim}
\alias{bind_rows_sim}
\title{Bind simulation list into dataframe}
\usage{
bind_rows_sim(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
bind_rows_sim(sim)
# ->
bind_rows(sim)
}

}
\keyword{internal}
