% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_situations.R
\name{cat_situations}
\alias{cat_situations}
\title{Format simulation data and observation data in order to consider all situations}
\usage{
cat_situations(list_sim = NULL, obs = NULL, obs_sd = NULL,
  force = TRUE, verbose = TRUE)
}
\arguments{
\item{list_sim}{A list (each element= version) of a list (each element= situation) of simulations \code{data.frame}s}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
A list of two : a list (each element= version) of a list of a single simulations \code{data.frame}
named "all_situations" and a list of a single observations \code{data.frame} named "all_situations"
}
\description{
Format simulation data and observation data in a list of a single data frame named "all_situations"
}
\keyword{internal}
