% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_situations.R
\name{cat_versions}
\alias{cat_versions}
\title{Format simulation and observation into a long data frame over versions}
\usage{
cat_versions(list_sim = NULL)
}
\arguments{
\item{list_sim}{A list (each element= version) of a list
(each element= situation) of simulations \code{data.frame}s}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s
(named by situation)}

\item{obs_sd}{A list (each element= situation) of \code{data.frame}s for the
standard deviation of the observations (named by situation)}
}
\value{
A dataframe
}
\description{
Format the list of simulation data and observation data
into a long data frame with a new column for the version
}
\keyword{internal}
