% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_situations.R
\name{cat_with_situation}
\alias{cat_with_situation}
\title{Format situations}
\usage{
cat_with_situation(dot_args, obs, obs_sd, all_situations, v_names)
}
\arguments{
\item{dot_args}{The dot arguments (the simulations).}

\item{obs}{The observations.}

\item{obs_sd}{The standard deviation of the observations.}

\item{all_situations}{Boolean. Do we need all situations treated
separately or as one.}

\item{v_names}{The version names.}
}
\value{
The categorized situations.
}
\description{
Format the data into a list of situations, with the situation repeated
as a column name, and the version also as a column name. In the case
of \code{all_situations==TRUE}, the situations are concatenated together
into one situation called "all_situations" (the data is a list of
one situation). The true situation name is still kept in the column
\code{sit_name} though.
}
\keyword{internal}
