% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stats.R
\name{check_stat_names}
\alias{check_stat_names}
\title{Check the names of the required statistics}
\usage{
check_stat_names(stat)
}
\arguments{
\item{stat}{A character vector of required statistics, or "all" for all.}
}
\value{
The names of the statistics to compute, without the unknown ones
(with a warning).
}
\description{
Check the names of the required statistics
}
\keyword{internal}
