% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_helpers.R
\name{create_plot}
\alias{create_plot}
\title{Generate a  plot}
\usage{
create_plot(
  data,
  x,
  y,
  geom_fun = ggplot2::geom_point,
  title = NULL,
  label = NULL,
  xlab = NULL,
  ylab = NULL,
  show.legend = TRUE,
  legend_colour = NULL,
  legend_shape = NULL,
  legend_size = NULL,
  add_geomArgs = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame of input data}

\item{x}{x coordiantes of the plotted data.}

\item{y}{y coordinates of the plotted data.}

\item{geom_fun}{A geometric function from the ggplot2 package}

\item{title}{Title of the plot.}

\item{label}{Vector of labels.}

\item{xlab}{label of the x-axis; passed on to ggplot's \code{xlab} function.}

\item{ylab}{label of the y-axis; passed on to ggplot's \code{xlab} function.}

\item{show.legend}{True by fault}

\item{legend_colour}{Title of the colour legend.}

\item{legend_shape}{Title of the shape legend.}

\item{legend_size}{Title of the size legend.}

\item{add_geomArgs}{List of addition arguments to \code{geom_fun}}

\item{...}{Additional arguments to be passed on to the ggplot's \code{geom_point} function.}
}
\value{
The graph of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a  plot applying different aestetics automatically.
}
\details{
Arguments \code{colour_legend}, \code{shape_legend} and \code{size_legend} only have an effect
if the respective aestetic is unsed, ie. if the \code{legend} argument is used or if
\code{aes(color=my_color)} or \code{aes(shape=my_shape)} is passed using \code{...}.
}
\examples{
\dontrun{
ToDo
}

}
