% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{detect_dataFrameLong}
\alias{detect_dataFrameLong}
\title{Detect whether a data frame is in long format}
\usage{
detect_dataFrameLong(data, dict)
}
\arguments{
\item{data}{A data frame}

\item{dict}{A named list}
}
\value{
A logic value indicating whether one column of \code{data} contains at least half the names of \code{dict}
}
\description{
Detect whether a data frame is in long format
}
