% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{detect_dataFrameWide}
\alias{detect_dataFrameWide}
\title{Detect whether a data frame is in wide format}
\usage{
detect_dataFrameWide(data, dict)
}
\arguments{
\item{data}{A data frame}

\item{dict}{A named list}
}
\value{
A logic value indicating whether at least half of the names of \code{list} are prefixes of column names of \code{data}.
}
\description{
Detect whether a data frame is in wide format
}
