% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{detect_item_case}
\alias{detect_item_case}
\title{Detect items cases}
\usage{
detect_item_case(
  type,
  is_mixture,
  one_version,
  has_distinct_situations,
  overlap
)
}
\arguments{
\item{type}{The type of plot required, either 'dynamic' or 'scatter}

\item{is_mixture}{A logical value indicating whether the crop is a mixture or not.}

\item{one_version}{A logical value indicating whether the plot has one or multiple versions (e.g. of the model).}

\item{has_distinct_situations}{A logical value indicating whether there are one or several situations to plot.}

\item{overlap}{A logical value indicating whether there is any overlapping variables in the plot.}
}
\value{
A unique character string for the plot.
}
\description{
This function returns a unique string based on the type of plot, and
whether the situation is a mixture or not, if there is one or multiple
versions to plot, and if there is one or several situations to plot
into the same plot.
The output is used to choose the right plotting function in a switch.
}
\keyword{internal}
