% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{detect_mixture}
\alias{detect_mixture}
\title{Detects if a situation is a mixture}
\usage{
detect_mixture(sim_situation)
}
\arguments{
\item{sim_situation}{A data frame containing the simulated data for
one situation.}
}
\value{
A logical value indicating if the situation is a mixture.
}
\description{
This function checks if the situation is a mixture based on
the presence of a column named "Dominance" and the uniqueness
of its values.
}
\examples{
sim_data <- data.frame(
  Dominance = c("Principal", "Principal", "Associated", "Associated")
)
detect_mixture(sim_data)
# Output: TRUE

sim_data <- data.frame(Dominance = c("Single Crop", "Single Crop"))
detect_mixture(sim_data)
# Output: FALSE

sim_data <- data.frame(lai = c(1, 1.2))
detect_mixture(sim_data)
# Output: FALSE

}
