% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{detect_mixture_version_overlap}
\alias{detect_mixture_version_overlap}
\title{Detect items cases}
\usage{
detect_mixture_version_overlap(is_mixture, one_version, overlap)
}
\arguments{
\item{is_mixture}{A logical value indicating whether the crop is a mixture or not.}

\item{one_version}{A logical value indicating whether the plot has one or multiple versions (e.g. of the model).}

\item{overlap}{A logical value indicating whether there is any overlapping variables in the plot.}
}
\value{
A character string indicating the case for computing the aesthetics of the plot.
}
\description{
This function detects the cases for computing the aesthetics of a plot based on
whether it is a mixture or not, whether it has one or multiple versions, and
whether there is any overlap.
}
\keyword{internal}
