% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{detect_mixture_version_situations}
\alias{detect_mixture_version_situations}
\title{Detect items cases for scatter plots}
\usage{
detect_mixture_version_situations(is_mixture, one_version, several_sit)
}
\arguments{
\item{is_mixture}{A logical value indicating whether the crop is a mixture or not.}

\item{one_version}{A logical value indicating whether the plot has one or multiple versions (e.g. of the model).}

\item{several_sit}{A logical value indicating whether there are one or several situations to plot.}
}
\value{
A character string indicating the case for computing the aesthetics of the plot.
}
\description{
This function detects the cases for computing the aesthetics of a plot based on
whether it is a mixture or not, whether it has one or multiple versions, and
whether there are one or several situations to plot into the same plot.
}
\keyword{internal}
