% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{detect_tibbleWithList}
\alias{detect_tibbleWithList}
\title{Detect whether a data frame is in tibble format}
\usage{
detect_tibbleWithList(data, dict)
}
\arguments{
\item{data}{A data frame}

\item{dict}{A named list}
}
\value{
A logic value indicating whether the \code{data} is in wide format and also contains at least on element of type \code{list}
}
\description{
Detect whether a data frame is in tibble format
}
