% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure}
\alias{ensure}
\title{Ensure the existence of a list of variables in a \code{cropr_input} object}
\usage{
ensure(data_object, parameters)
}
\arguments{
\item{data_object}{A \code{cropr_input} object}

\item{parameters}{A list of characters}
}
\value{
A list of three elements:
\enumerate{
\item $object, a \code{cropr_input} object containing all variables in \code{parameters} if they could be calculated from the given \code{data_object}
\item $missing, a list keeping track of missing variables that could be provided to calculate all parameters.
The format of the list allows it to be printed by the \code{print_missingTree} function.
\item $success, a logical vector indicating whether the parameters could be calculated.
}
}
\description{
Ensure the existence of a list of variables in a \code{cropr_input} object
}
\details{
The function scans the package for functions of the form "ensure_\emph{parameter_name}".
Only functions found by this scan will be used to calculate new parameters.
}
