% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R, R/ensure_soil.R, R/ensure_weather.R
\name{ensure_nb_below_0}
\alias{ensure_nb_below_0}
\alias{ensure_nb_above_35}
\alias{ensure_saturated_wtr_cap}
\alias{ensure_layer_saturated_wtr_cap}
\alias{ensure_depth}
\alias{ensure_summary_year}
\alias{ensure_summary_station_name}
\alias{ensure_rainfall_cumulated}
\alias{ensure_temp_day_mean}
\alias{ensure_temp_mean}
\title{Ensure the existence of a variable}
\usage{
ensure_nb_below_0(weather)

ensure_nb_above_35(weather)

ensure_saturated_wtr_cap(soil)

ensure_layer_saturated_wtr_cap(soil)

ensure_depth(soil)

ensure_summary_year(weather)

ensure_summary_station_name(weather)

ensure_rainfall_cumulated(weather)

ensure_temp_day_mean(weather)

ensure_temp_mean(weather)
}
\arguments{
\item{weather}{A \code{cropr_input} object containing weather data}

\item{soil}{A \code{cropr_input} object containing soil data}
}
\value{
A list of three elements:
\enumerate{
\item $object, a \code{cropr_input} object containing the variable in the function name if it could be calculated from the given data object
\item $missing, a list keeping track of missing variables that could be provided to calculate the variable.
The format of the list allows it to be printed by the \code{print_missingTree} function.
\item $success, a logical value indicating whether the variable could be calculated.
}
}
\description{
Ensure the existence of a variable
}
