% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_wrapper}
\alias{ensure_wrapper}
\title{Ensure the existence of a list of variables in a \code{cropr_input} object}
\usage{
ensure_wrapper(data_object, parameters, type)
}
\arguments{
\item{data_object}{A \code{cropr_input} object}

\item{parameters}{A list of characters}

\item{type}{A character}
}
\value{
A \code{cropr_input} object containing all variables in \code{parameters}.
}
\description{
Ensure the existence of a list of variables in a \code{cropr_input} object
}
\details{
This function will throw an error if any of the parameters can not be calculated with the given \code{data_object}.

The \code{type} argument is ony used to make the error message more specific.
}
