% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plot.R
\name{extract_plot}
\alias{extract_plot}
\title{Extract plot(s) from ggplot}
\usage{
extract_plot(
  plot,
  var = NULL,
  situation = NULL,
  force = TRUE,
  verbose = TRUE,
  situations = lifecycle::deprecated()
)
}
\arguments{
\item{plot}{The output of plot_situations.}

\item{var}{Vector of variable names for which plots have to be extracted. Optional, all variables considered by default.}

\item{situation}{A list of situations names to extract from a list of ggplots.}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Logical value for displaying information while running.}

\item{situations}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{situations} is no
longer supported, use \code{situation} instead.}
}
\value{
A (printed) list of ggplot objects, each element being a plot for a situation
}
\description{
Extract a plot corresponding to one or several variables from a ggplot.
}
