% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{find_indexInList}
\alias{find_indexInList}
\title{Get the indexes of all elements of a list that contain a certain value}
\usage{
find_indexInList(value, list)
}
\arguments{
\item{value}{A value}

\item{list}{A list of lists}
}
\value{
The indexes all elements of \code{list} that contain value
}
\description{
Get the indexes of all elements of a list that contain a certain value
}
