% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{find_nameInList}
\alias{find_nameInList}
\title{Get the names of all elements of a list that contain a certain value}
\usage{
find_nameInList(value, list)
}
\arguments{
\item{value}{A value}

\item{list}{A list of lists}
}
\value{
The names all elements of \code{list} that contain value
}
\description{
Get the names of all elements of a list that contain a certain value
}
