% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_formatting.R
\name{format_cropr}
\alias{format_cropr}
\title{Format simulations and observations from CropR format to a format usable by
CroplotR}
\usage{
format_cropr(
  sim,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  successive = NULL,
  reference_var = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sim}{A simulation data.frame}

\item{obs}{An observation data.frame}

\item{obs_sd}{A data.frame with observation standard deviations}

\item{type}{The type of plot required, either "dynamic" or "scatter"}

\item{select_dyn}{Which data to plot when \code{type= "dynamic"}? See details.}

\item{select_scat}{Which data to plot when \code{type= "scatter"}? See details.}

\item{successive}{A list of lists containing the situations to be represented
as a contiguous sequence when type = "dynamic" (dates should be contiguous)
when \code{type = "dynamic"} (implies that the situations are correctly ordered).}

\item{reference_var}{Variable selected on x-axis when type is scatter and
select_scat is res. It is possible to select between observation and
simulation of the reference variable.}

\item{verbose}{Logical value for displaying information while running.}
}
\value{
A pre-formatted \code{data.frame} or \code{NULL} if the formatting is not
possible (e.g. type="scatter" but no common variables in obs and sim).
}
\description{
Format simulations (and observations if any) for plotting. This
function can be used as a template to include other models in CroPlotR.
}
\details{
The \code{select_dyn} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and
observations when there are some
\item "common": variables with simulations outputs and observations in common
(used when \code{type= "scatter"} )
\item "obs": all variables with observations, and simulations outputs when there
are some
\item "all": all variables with any observations or simulations outputs
}

The \code{select_scat} argument can be:
\itemize{
\item "sim" (the default): plots observations in X and simulations in Y.
\item "res": plots observations in X and residuals(observations-simulations)in Y.
}
}
\examples{
\dontrun{
# remotes::install_github("SticsRPacks/SticsRPacks")
workspace <- system.file(file.path("extdata", "stics_example_1"),
  package = "CroPlotR"
)
situation <- SticsRFiles::get_usms_list(
  file =
    file.path(workspace, "usms.xml")
)[1]
sim <- SticsRFiles::get_sim(workspace = workspace, usm = situation)
obs <- SticsRFiles::get_obs(workspace = workspace, usm = situation)
formated_df <- format_cropr(
  sim$`IC_Wheat_Pea_2005-2006_N0`,
  obs$`IC_Wheat_Pea_2005-2006_N0`
)
options(max.print = 100)
formated_df
}
}
