% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_formating.R
\name{format_stics}
\alias{format_stics}
\title{Format a STICS simulation (and observation)}
\usage{
format_stics(sim, obs = NULL, plot = c("sim", "common", "obs", "all"))
}
\arguments{
\item{sim}{A simulation list of data.frames}

\item{obs}{An observation list of data.frames}

\item{plot}{The priority to either simulation or observation points if missing values (see details)}
}
\value{
A list of two: a pre-formatted \code{data.frame}, and a colouring expression, or \code{NULL} if
the formatting is not possible (e.g. plot="common" but no common variables in obs and sim).
}
\description{
Format a STICS simulation (and observation if any) for plotting. This
function can be used as a template to include other models in CroPlotR.
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim": all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with obervations, and simulations outputs when there are some
\item "all": all variables with any obervations or simulations outputs
}
}
\examples{
workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situation= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))[1]
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situation)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situation)
formated_df= format_stics(sim$`IC_Wheat_Pea_2005-2006_N0`,obs$`IC_Wheat_Pea_2005-2006_N0`)
options(max.print= 100)
formated_df
}
