% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_formatting.R
\name{generic_formatting}
\alias{generic_formatting}
\title{Generic formatting}
\usage{
generic_formatting(
  df,
  var,
  overlap,
  situation_group,
  type,
  shape_sit,
  several_sit,
  total_vers,
  num_vers
)
}
\arguments{
\item{df}{A data frame to be formatted (e.g. output from \code{format_cropr}).}

\item{var}{A character string indicating the variable to be formatted.}

\item{overlap}{A logical value indicating whether there is any overlapping
variables desired in the plot.}

\item{situation_group}{A character string indicating the situation group.}

\item{type}{A character string indicating the type of plot (dynamic or
scatter).}

\item{shape_sit}{A logical value indicating whether the shape should be
based on the situation.}

\item{several_sit}{A logical value indicating whether there are several
situations to plot.}

\item{total_vers}{An integer indicating the total number of versions.}

\item{num_vers}{An integer indicating the id of the current version.}
}
\value{
A formatted data frame.
}
\description{
This function performs generic formatting on a data frame based on the input
parameters.
}
