% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_helpers.R
\name{geomTextRepel<-}
\alias{geomTextRepel<-}
\title{Replace the plot's \code{GeomTextRepel} object}
\usage{
geomTextRepel(plot) <- value
}
\arguments{
\item{plot}{A plot}

\item{value}{A new \code{GeomTextRepel} object}
}
\value{
The plot \code{plot} whose \code{GeomTextRepel} object has been replaced by \code{value}
}
\description{
Replace the plot's \code{GeomTextRepel} object
}
\examples{
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
soil_data_wide <- readRDS(file.path(workspace, "soil_data_wide.rds"))

soil <- set_soil(
    soil_data_wide,
    id="name",
    layer_thickness=list("epc", "cm"),
    layer_water_field_cap=list("HCCF", "g/g"),
    layer_water_wilting_pt=list("HMINF", "g/g"),
    layer_bulk_density=list("DAF", "g/cm^3"),
    organic_N_conc=list("norg", "g/g")
)

plot <- plot_soil(soil, type="thickness.mswc")
geomTextRepel(plot) <- ggrepel::geom_text_repel(max.overlaps = Inf)

}
