% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{get_indVar}
\alias{get_indVar}
\title{Get the column of a data frame that contains at least half of the names of a list.}
\usage{
get_indVar(data, dict)
}
\arguments{
\item{data}{A data frame}

\item{dict}{A named list}
}
\value{
The index of the column in \code{data} that contains most of the names of \code{dict}. Returns \code{character(0)} if
to column contains more than half of the names.
}
\description{
Get the column of a data frame that contains at least half of the names of a list.
}
