% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plotting_scatter.R
\name{give_reference_var}
\alias{give_reference_var}
\title{Get reference variable for plotting}
\usage{
give_reference_var(reference_var)
}
\arguments{
\item{reference_var}{The reference variable name, if NULL "Observed" is used}
}
\value{
A list with two elements:
\itemize{
\item reference_var: The reference variable name: "Observed" or "Reference"
\item reference_var_name: The display name for the reference variable
}
}
\description{
Return the reference variable and its display name for scatter
plots
}
\keyword{internal}
