% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plotting_scatter.R
\name{give_y_var_type}
\alias{give_y_var_type}
\title{Get y variable type for plotting}
\usage{
give_y_var_type(select_scat)
}
\arguments{
\item{select_scat}{Selection type, either "sim" for Simulated or any other
value for Residuals}
}
\value{
A character string, either "Simulated" or "Residuals"
}
\description{
Return the type of y variable for scatter plots based on
selection
}
\keyword{internal}
