% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{is_prefixOf}
\alias{is_prefixOf}
\title{Check if one character is a prefix of another}
\usage{
is_prefixOf(x, vec_names)
}
\arguments{
\item{x}{character}

\item{vec_names}{A vector of characters}
}
\value{
A logic vector indicating whether \code{x} is a prefix of every element in \code{vec_names}
}
\description{
Check if one character is a prefix of another
}
