% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_plot_situations_arguments.R
\name{parse_plot_situations_args}
\alias{parse_plot_situations_args}
\title{Parse plot situations arguments}
\usage{
parse_plot_situations_args(
  dot_args,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = NULL,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  successive = NULL,
  shape_sit = c("none", "txt", "symbol", "group"),
  situation_group = NULL,
  reference_var = NULL,
  force = FALSE,
  verbose = TRUE
)
}
\value{
A list of parsed arguments, eventually corrected according to the
compatibility between options.
}
\description{
This function parses the arguments for \code{plot_situations}.
}
\keyword{internal}
