% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting.R
\name{plot.cropr_input}
\alias{plot.cropr_input}
\title{Generate all possible plots based on any set of characteristics}
\usage{
\method{plot}{cropr_input}(...)
}
\arguments{
\item{...}{ToDo}
}
\value{
The plot of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a plot of type \code{type} reflecting soil charactersitics of
the soil data object \code{soil}, possibly complemented by information from other
data objects (eg. weather, situation).
}
\details{
Use the \code{\link{data_soil}}, \code{{data_weather}} and \code{{data_situation}}
functions to create respective data objects from user-given data.
}
\examples{
\dontrun{
# ToDo
}


}
