% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ploting.R
\name{plot.statistics}
\alias{plot.statistics}
\title{Plot statistics}
\usage{
\method{plot}{statistics}(
  x,
  xvar = c("group", "situation"),
  type = c("bar", "radar"),
  crit_radar = NULL,
  title = NULL,
  force = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{The output of \code{\link[=summary.stics_simulation]{summary.stics_simulation()}}}

\item{xvar}{The variable to use in x, either the group or the situation (the other is used for colouring)}

\item{title}{The plot title}

\item{...}{Other arguments to pass (for backward compatibility only)}
}
\value{
Return a ggplot object with statistics
}
\description{
Plot statistics
}
\examples{
# Importing an example with three situations with observation:
workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situations= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situations)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situations)

# R2 and nRMSE stats for the simulation:
stats= summary(sim, obs= obs, stat= c("R2","nRMSE"))
plot(stats)

# Change the group name:
stats= summary("stics v9.0"= sim, obs= obs, stat= c("R2","nRMSE"))
plot(stats)

# R2 and nRMSE stats for two groups of simulations:
summary(sim1= sim, sim2= sim, obs=obs, stat= c("R2","nRMSE"))

}
