% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_graphes.R
\name{plot__thickness.mswc}
\alias{plot__thickness.mswc}
\alias{plot__limiting.temperatures}
\title{Generate a graph of specified type}
\usage{
plot__thickness.mswc(soil, histogram, ...)

plot__limiting.temperatures(weather, histogram, ...)
}
\arguments{
\item{soil}{A \code{cropr_input} object containing soil data}

\item{weather}{A \code{cropr_input} object containing weather data}
}
\value{
A graph created by the \code{ggplot2} package
}
\description{
Generate a graph of specified type
}
\details{
The function names of the form 'plot_\emph{plot_type}' are required for these specific plot functions
to be found by the \code{plot_generic_input} function
}
