% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plotting - Copie.R,
%   R/generic_plotting.R
\name{plot_generic_situation}
\alias{plot_generic_situation}
\title{Generic plotting of a situation}
\usage{
plot_generic_situation(
  sim,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = var,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  successive = NULL,
  shape_sit = c("none", "txt", "symbol", "group"),
  situation_group = NULL,
  total_vers = 1,
  num_vers = 1,
  reference_var = NULL,
  force = TRUE,
  verbose = TRUE,
  formater
)

plot_generic_situation(
  sim,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = var,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  successive = NULL,
  shape_sit = c("none", "txt", "symbol", "group"),
  situation_group = NULL,
  total_vers = 1,
  num_vers = 1,
  reference_var = NULL,
  force = TRUE,
  verbose = TRUE,
  formater
)
}
\arguments{
\item{sim}{A simulation data.frame}

\item{obs}{An observation data.frame (variable names must match)}

\item{obs_sd}{A data.frame of standard deviations of observations}

\item{select_dyn}{Which data to plot when \code{type= "dynamic"}? See details.}

\item{select_scat}{Which data to plot when \code{type= "scatter"}? See details.}

\item{var}{A vector of variables that should be displayed on the graph.}

\item{title}{The plot title}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, plot all
situations on the same graph.}

\item{overlap}{A list of lists containing the variables to represent on
the same graph when \code{type = "dynamic"}.}

\item{successive}{A list of lists containing the situations to be represented
as a contiguous sequence when \code{type = "dynamic"}
(implies that the situations are correctly ordered).}

\item{shape_sit}{Shape to differentiate between situations when
\code{all_situations= TRUE}. See details.}

\item{situation_group}{A list of lists of situations to gather when
\code{shape_sit= "group"}.}

\item{reference_var}{Variable selected on x-axis when type is scatter and
select_scat is res. It is possible to select
between observation and simulation of the reference variable.
(examples : reference_var = "lai_n_obs", reference_var = "mafruit_sim")}

\item{force}{Continue if the plot is not possible ? E.g. no observations for
scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}

\item{formater}{The function used to format the models outputs and
observations in a standard way. You can design your own function that format
one situation and provide it here.}
}
\value{
A ggplot object

A ggplot object
}
\description{
Plots outputs of a model (and observations) for one situation.
This function is used as a generic plotting function for any models.
To use it with your own model, please provide a wrapper function around your
model to format the outputs used by this function (see \code{\link[=format_cropr]{format_cropr()}} for a
template), and then provide your custom function as an argument to this one.

Plots outputs of a model (and observations) for one situation.
This function is used as a generic plotting function for any models.
To use it with your own model, please provide a wrapper function around your
model to format the outputs used by this function (see \code{\link[=format_cropr]{format_cropr()}} for a
template), and then provide your custom function as an argument to this one.
}
\details{
The \code{select_dyn} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and
observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with observations, and simulations outputs
when there are some
\item "all": all variables with any observations or simulations outputs
}

The \code{select_scat} argument can be:
\itemize{
\item "sim" (the default): plots observations in X and simulations in Y.
\item "res": plots observations in X and residuals(observations-simulations)in Y.
}

The \code{shape_sit} argument can be:
\itemize{
\item "none" (the default): Same shape for all situations.
\item "txt": Writes the name of the situation above each point.
\item "symbol": One shape for each situation.
\item "group": One shape for each group of situations described in
\code{situation_group}.
}

The \code{select_dyn} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and
observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with observations, and simulations outputs
when there are some
\item "all": all variables with any observations or simulations outputs
}

The \code{select_scat} argument can be:
\itemize{
\item "sim" (the default): plots observations in X and simulations in Y.
\item "res": plots observations in X and residuals(observations-simulations)in Y.
}

The \code{shape_sit} argument can be:
\itemize{
\item "none" (the default): Same shape for all situations.
\item "txt": Writes the name of the situation above each point.
\item "symbol": One shape for each situation.
\item "group": One shape for each group of situations described in
\code{situation_group}.
}
}
\note{
The error bar will be equal to 2*\code{obs_sd} on each side of the point to
have 95\% confidence.

The error bar will be equal to 2*\code{obs_sd} on each side of the point to
have 95\% confidence.
}
\keyword{internal}
