% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_graphes.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generate a scatter plot}
\usage{
plot_scatter(
  data,
  x,
  y,
  geom_fun = ggplot2::geom_point,
  title = NULL,
  label = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_colour = NULL,
  legend_shape = NULL,
  legend_size = NULL,
  add_geomArgs = NULL,
  ...
)
}
\arguments{
\item{x}{x coordiantes of the plotted data.}

\item{y}{y coordinates of the plotted data.}

\item{title}{Title of the scatter plot.}

\item{label}{Vector of labels for the scatterd points.}

\item{xlab}{label of the x-axis; passed on to ggplot's \code{xlab} function.}

\item{ylab}{label of the y-axis; passed on to ggplot's \code{xlab} function.}

\item{...}{Additional arguments to be passed on to the ggplot's \code{geom_point} function.}

\item{max.overlaps}{Exclude text labels that overlap too many things. Defaults to Inf.}

\item{colour.legend}{Title of the colour legend.}

\item{shape.legend}{Title of the shape legend.}

\item{size.legend}{Title of the size legend.}
}
\value{
The graph of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a scatter plot applying different aestetics automatically.
}
\details{
Arguments \code{colour_legend}, \code{shape_legend} and \code{size_legend} only have an effect
if the respective aestetic is unsed, ie. if the \code{legend} argument is used or if
\code{aes(color=my_color)} or \code{aes(shape=my_shape)} is passed using \code{...}.
}
\examples{
\dontrun{
ToDo
}

}
