% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ploting.R
\name{plot_situations}
\alias{plot_situations}
\title{Generic plotting function for all models}
\usage{
plot_situations(
  ...,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "all"),
  Title = NULL,
  plot_function = plot_stics_situation
)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{type}{The type of plot requested, either "dynamic" (date in X, variable in Y) or scatter (simulated VS observed)}

\item{plot}{Which data to plot in priority when \code{type= "dynamic"}? See details.}

\item{Title}{A vector of plot titles, named by situation. Use the situation name if \code{NULL}, recycled if length one.}

\item{plot_function}{The function used to make the plots. Please design your own function that plots one situation and provide
it as an argument of this function.}
}
\value{
A (printed) list of ggplot objects, each element being a plot for a situation
}
\description{
simulation outputs for one or several situations with or without observations, grouped
by a model version (or any group actually). Please use this function as a wrapper for your
own model function that you can design based on \code{\link[=plot_stics_situation]{plot_stics_situation()}}. Note that all arguments should be the
same.
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common (used when \code{type= "scatter"} )
\item "obs": all variables with obervations, and simulations outputs when there are some
\item "all": all variables with any obervations or simulations outputs
}
}
\note{
The plots titles are given by their situation name.
}
