% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_graphes.R
\name{plot_thickness.mswc.norg}
\alias{plot_thickness.mswc.norg}
\title{Generate a graph: thickness, maximum  water content and organic nitrogen per soil}
\usage{
plot_thickness.mswc.norg(soil, interactive = FALSE, ...)
}
\arguments{
\item{name}{Name of each soil.}

\item{thickness}{Thickness of each soil.}

\item{mswc}{Maximum water content of each soil.}

\item{norg}{Amount og organic nitrogen in each soil.}

\item{max.overlaps}{Exclude names of certain soils that would overlap too many things.
Defaults to Inf meaning that all names are shown.}
}
\value{
The respective graph as \code{ggplot2} object.
}
\description{
Generates a graph decpicting the tickness, the maximum water content and the organis nitrogen per soil.
}
\examples{
\dontrun{
ToDo
}

}
