% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_plotting.R
\name{plot_usm}
\alias{plot_usm}
\title{Basic function for plotting}
\usage{
plot_usm(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "all"),
  Title = NULL
)
}
\arguments{
\item{sim}{A simulation list of data.frames}

\item{obs}{An observation list of data.frames}

\item{type}{The type of plot required, either "dynamic" or "scatter"}

\item{plot}{The priority to either simulation or observation points if missing values (see details)}

\item{Title}{The plot title}
}
\value{
A ggplot object
}
\description{
Basic function for plotting
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim": all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with obervations, and simulations outputs when there are some
\item "all": all variables with any obervations or simulations outputs
}
}
\keyword{internal}
