% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_plotting.R
\name{plot_usms}
\alias{plot_usms}
\title{Plot usms}
\usage{
plot_usms(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "all"),
  Title = NULL
)
}
\arguments{
\item{sim}{A list of simulation output \code{data.frame}s for each situation (named by situation)}

\item{obs}{A list of observation \code{data.frame}s for each situation (named by situation)}

\item{type}{The type of plot, either dynamic (Date in X, variable in Y) or scatter (Simulated VS observed)}

\item{plot}{Which data to plot ? See details.}

\item{Title}{The plots title. Gives the situation as Title if NULL.}
}
\value{
A list of ggplot, each element being a plot for a situation
}
\description{
simulation outputs for one or several USMs with or without observations
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim": all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with obervations, and simulations outputs when there are some
\item "all": all variables with any obervations or simulations outputs
}
}
