% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{print_missingTree}
\alias{print_missingTree}
\title{Print a list as returned by the \code{ensure} function's $missing return}
\usage{
print_missingTree(missing, level = 0)
}
\arguments{
\item{missing}{A list}

\item{level}{Level of indention}
}
\value{
A string that prints \code{missing} as a AND/OR tree in a very concise way
}
\description{
Print a list as returned by the \code{ensure} function's $missing return
}
