% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save CroPlotR plots}
\usage{
save_plot(plot, path, filename = "Graphs", main = "Plots",
  file_per_var = FALSE, stats = NULL, force = TRUE, verbose = TRUE)
}
\arguments{
\item{plot}{A list of ggplots : output of \code{plot_situations()}}

\item{path}{The path to the directory where to save the plots}

\item{filename}{Name of the pdf file}

\item{main}{Main title of the pdf document}

\item{file_per_var}{If \code{TRUE}, produces one file per variable instead of one with all variables inside}

\item{stats}{Output of \code{statistics_situations}with \code{all_situations = FALSE}. Needed if \code{file_per_var} is TRUE.
It is used to classify situations according to the descending RMSE.}

\item{force}{Continue if the plot is not possible ? If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
Save the plots in a pdf file in the folder specified by the \code{path}
}
\description{
Save the plots to a pdf file
}
