% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_data.R
\name{set_soil}
\alias{set_soil}
\title{Collect soil characteristics for further use in other CorPlotR-functions}
\usage{
set_soil(x, ...)
}
\arguments{
\item{data}{A data.frame or list containing soil caracteristics}

\item{name}{Soil name}

\item{thickness}{Soil thickness (sum of all layers).}

\item{mswc}{Soil maximum soil water content (sum of all layers).}

\item{norg}{Soil organic nitrogen content.}

\item{var5}{Some other parameter.}
}
\value{
A list containing the given variables with fixed CroPlotR names of class \code{cropr_input}.
}
\description{
Transforms user-provided soil data into an object that can be read by other CroPlotR
functions, most notably by the plot functions. Varibles can either be given explicitly
(eg. var1 = 5) or through a name refering to a variable in data (eg. var2 = mswc, where
data$mswc exists).
}
\examples{
\dontrun{
# ToDo: add data that makes this example work
workspace= "path_to_workspace"

soil_data <- SticsRFiles::get_param_xml(file.path(workspace, "sols.xml"))[[1]]
soil_data$name <- SticsRFiles::get_soils_list(file.path(workspace, "sols.xml"))

soil_data$MSWC <- soil_data$epc*(soil_data$HCCF-soil_data$HMINF)
soil_data$MSWC <- colMeans(matrix(soil_data$MSWC, nrow=5))
soil_data$thickness <- colSums(matrix(soil_data$epc, nrow=5))
soil <- data_soil(data = soil_data, thickness=thickness, mswc=MSWC, norg=norg, name=name)
}

}
