% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_weather.R
\name{set_weather}
\alias{set_weather}
\title{Collect data on weather variables for further use in plot-functions}
\usage{
set_weather(
  data,
  id = NULL,
  station_name = NULL,
  temp_day_max = NULL,
  temp_day_min = NULL,
  temp_day_mean = NULL,
  rainfall_day = NULL,
  radiation_day = NULL,
  etp_day = NULL,
  year = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame containing weather data}

\item{id}{Unique weather identification}

\item{station_name}{Weather station name}

\item{temp_day_max}{Air temperature, daily maximum}

\item{temp_day_min}{Air temperature, daily minimum}

\item{temp_day_mean}{Air temperature, daily mean}

\item{rainfall_day}{Rainfall, per day}

\item{radiation_day}{Radiation, per day}

\item{etp_day}{Potential_evapotranspiration, per day}

\item{year}{year of observation}

\item{verbose}{Provide extra information about the function's inner procedures?}
}
\value{
A list of class \code{cropr_input} containing all necessary information for plotting.
}
\description{
Transforms user-provided weather data into an object that can be read by other CroPlotR
functions, most notably by the plot functions. The data on all variables has to be given through
a single data frame. The names of every variable and its unit have to be specified.
}
\details{
The column name in \code{data} and the unit of every variable have to be given as a list of two characters.
}
