% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stats.R
\name{situation_pairs}
\alias{situation_pairs}
\title{Observed / simulated pairs for one situation}
\usage{
situation_pairs(sim, obs, verbose = TRUE)
}
\arguments{
\item{sim}{A simulation data.frame, possibly with a \code{version} column}

\item{obs}{An observation data.frame (variable names must match)}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
A data.frame with columns (version), Plant, variable, Observed and
Simulated, or \code{NULL} if there are no common variables between \code{sim} and
\code{obs}.
}
\description{
Observed / simulated pairs for one situation
}
\keyword{internal}
