% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure - Copie.R, R/ensure.R, R/ensure_soil.R,
%   R/ensure_weather - Copie.R, R/ensure_weather.R
\name{specific_ensure_doc}
\alias{specific_ensure_doc}
\alias{ensure_soil_max_wtr_cap}
\alias{ensure_layer_max_wtr_cap}
\alias{ensure_thickness}
\alias{ensure_summary_year}
\alias{ensure_summary_station_name}
\alias{ensure_rainfall_cumulated}
\alias{ensure_temp_day_mean}
\alias{ensure_temp_mean}
\alias{ensure_nb_below_0}
\alias{ensure_nb_above_35}
\alias{ensure_nb_below_threshold_Tmin}
\alias{ensure_nb_above_threshold_Tmax}
\title{Ensure the existence of a variable}
\usage{
ensure_soil_max_wtr_cap(soil)

ensure_layer_max_wtr_cap(soil)

ensure_thickness(soil)

ensure_summary_year(weather)

ensure_summary_station_name(weather)

ensure_rainfall_cumulated(weather)

ensure_temp_day_mean(weather)

ensure_temp_mean(weather)

ensure_nb_below_0(weather)

ensure_nb_above_35(weather)

ensure_summary_year(weather)

ensure_summary_station_name(weather)

ensure_rainfall_cumulated(weather)

ensure_temp_day_mean(weather)

ensure_temp_mean(weather)

ensure_nb_below_threshold_Tmin(weather, threshold_Tmin)

ensure_nb_above_threshold_Tmax(weather, threshold_Tmax)
}
\arguments{
\item{soil}{A \code{cropr_input} object containing soil data}

\item{weather}{A \code{cropr_input} object containing weather data}
}
\value{
A list of three elements:
\enumerate{
\item $object, a \code{cropr_input} object containing the variable in the function name if it could be calculated from the given data object
\item $missing, a list keeping track of missing variables that could be provided to calculate the variable.
The format of the list allows it to be printed by the \code{print_missingTree} function.
\item $success, a logical value indicating whether the variable could be calculated.
}

A list of three elements:
\enumerate{
\item $object, a \code{cropr_input} object containing the variable in the function name if it could be calculated from the given data object
\item $missing, a list keeping track of missing variables that could be provided to calculate the variable.
The format of the list allows it to be printed by the \code{print_missingTree} function.
\item $success, a logical value indicating whether the variable could be calculated.
}
}
\description{
Ensure the existence of a variable

Ensure the existence of a variable
}
\keyword{internal}
