% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stats.R
\name{stats_description}
\alias{stats_description}
\title{Description of the available statistics}
\usage{
stats_description()
}
\value{
A one row data.frame with the description of each statistic, named
by statistic.
}
\description{
Description of the available statistics
}
\keyword{internal}
