% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subst_parenth.R
\name{subst_parenth}
\alias{subst_parenth}
\title{Substitute "(xxx)" by *_xxx"}
\usage{
subst_parenth(var_names)
}
\arguments{
\item{var_names}{Vector of strings to transform}
}
\value{
Vector of transformed strings
}
\description{
Substitute "(xxx)" by *_xxx" in case the user give names with parenthesis
while they are transformed by R in _
}
\examples{
var_names <- c("var1","var2(n)")
CroPlotR:::subst_parenth(var_names)

}
\keyword{internal}
