% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_stats.R
\name{summary.stics_simulation}
\alias{summary.stics_simulation}
\title{Summary statistics of simulations}
\usage{
\method{summary}{stics_simulation}(..., obs)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here (see \code{\link[=statistics]{statistics()}} and \code{\link[=format_stics]{format_stics()}} for more information).}
}
\value{
A list of statistics \code{data.frame}s named by situation
}
\description{
Summary statistics for one or several situations with observations, eventually grouped
by a model version (or any group actually)
}
\examples{
\dontrun{
# Importing an example with three situations with observation:
workspace= system.file(file.path("extdata", "STICS"), package = "CroPlotR")
situations= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situations)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situations)

summary(sim,obs=obs)
}
}
