% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_helpers.R
\name{symbols_applyIfClass}
\alias{symbols_applyIfClass}
\title{Add function call to symbols/quosures in nested list whose evaluation in some data is of a given class}
\usage{
symbols_applyIfClass(data, geom_args, funName, className)
}
\arguments{
\item{data}{A data frame}

\item{geom_args}{A list of symbols or quosures}

\item{funName}{Name of the function to apply}

\item{className}{Name of a class}
}
\value{
The list \code{geom_args} where the function \code{funName} is applied to all symbols whose evaluation in
\code{data} is of class \code{className}.
}
\description{
Add function call to symbols/quosures in nested list whose evaluation in some data is of a given class
}
