% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{template_aesthetics}
\alias{template_aesthetics}
\title{Template aesthetics}
\usage{
template_aesthetics()
}
\value{
A list of aesthetics for a plot.
}
\description{
This function returns a template list of aesthetics for a plot, including
color, shape, linetype, and group.
}
\keyword{internal}
