% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotType_doc.R
\name{thickness.mswc.norg}
\alias{thickness.mswc.norg}
\title{Plot type thickness.mswc.norg}
\arguments{
\item{name}{?}

\item{thickness}{?}

\item{mswc}{?}

\item{norg}{?}

\item{max.overlaps}{?}
}
\description{
Plot type thickness.mswc.norg
}
