% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotType_doc.R
\name{thickness.mswc}
\alias{thickness.mswc}
\title{Plot type thickness.mswc}
\arguments{
\item{depth}{Depth of the entire soil}

\item{saturated_wtr_cap}{Saturated water capacity of the entire soil}

\item{organic_N_conc}{Organic nitrogen concentration in the soil}
}
\description{
Plot type thickness.mswc
}
