% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_data.R
\name{variables_extract_from_lists}
\alias{variables_extract_from_lists}
\title{Extract variables from multiple lists}
\usage{
variables_extract_from_lists(vars, ...)
}
\arguments{
\item{vars}{A list, potentially containing variable names.}

\item{...}{Any number of lists.}
}
\value{
The list \code{vars} where every name that can be
matched to a variable in any of the lists in \code{...} is replaced this varaible's
value.
}
\description{
For every name that corresponds to a variable in the data,
replaces the variable name in \code{vars} by its value in data.
}
\examples{
\dontrun{
vars <- list("var1"=quote(a), "var2"=quote(c))
soil <- list(a=0.5, b=2, c=5)
variables_extract_from_lists(vars, soil)
}

}
