% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{aesthetics}
\alias{aesthetics}
\title{Manages the aesthetics of the graphics}
\usage{
aesthetics(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  overlap = NULL,
  several_sit = FALSE,
  shape_sit = c("none", "txt", "symbol", "group"),
  one_version = TRUE,
  iversion = 1,
  dot_args = NULL
)
}
\arguments{
\item{sim}{A simulation list of data.frames}

\item{obs}{An observation list of data.frames}

\item{type}{The type of plot required, either "dynamic" or "scatter"}

\item{overlap}{A list of lists containing the variables to represent on the
same graph when \code{type = "dynamic"}}

\item{several_sit}{Boolean. Must be equal to \code{TRUE} if sim and obs gather
more than one situation and if situations should be differentiated
on the graph.}

\item{shape_sit}{Shape to differentiate between situations when
\code{all_situations= TRUE}. See details.}

\item{one_version}{Boolean. Must be \code{TRUE} if several versions will be
plotted on the same graph.}

\item{iversion}{Integer. Version number of sim}

\item{dot_args}{List of dot arguments given to plot function}
}
\value{
An aesthetics expression which will allow to manage the aesthetics
of the graphics
}
\description{
Manages the shape, color and line type of the graphics according
to their content
}
\details{
The \code{shape_sit} argument can be:
\itemize{
\item "none" (the default): Same shape for all situations.
\item "txt": Writes the name of the situation above each point.
\item "symbol": One shape for each situation.
\item "group": One shape for each group of situations described in
\code{situation_group}.
}
}
\keyword{internal}
