% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics_dynamic.R
\name{aesthetics_dynamic}
\alias{aesthetics_dynamic}
\title{Manages the aesthetics of the graphics for dynamic plots}
\usage{
aesthetics_dynamic(
  sim,
  aesthetics = template_aesthetics(),
  overlap = NULL,
  one_version = TRUE,
  iVersion = 1,
  dot_args = NULL,
  is_mixture = FALSE
)
}
\arguments{
\item{sim}{A simulation list of data.frames}

\item{aesthetics}{Template aesthetics list}

\item{overlap}{A list of lists containing the variables to represent on the
same graph when \code{type = "dynamic"}}

\item{one_version}{Boolean. Must be \code{TRUE} if several versions will be
plotted on the same graph.}

\item{iVersion}{Integer. Version number of sim}

\item{dot_args}{List of dot arguments given to plot function}

\item{is_mixture}{Boolean indicating whether the crop is a mixture or not}
}
\value{
An aesthetics expression which will allow to manage the aesthetics
of the graphics
}
\description{
Manages the shape, color and line type of the graphics according
to their content
}
\keyword{internal}
