% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_bind_split.R
\name{bind_rows_sim}
\alias{bind_rows_sim}
\title{Bind rows for class Cropr}
\usage{
bind_rows_sim(sim)
}
\arguments{
\item{sim}{Simulation outputs in Cropr format (named list of \code{data.frame} for each situation,
having the attribute cropr_simulation)}
}
\value{
A single data.frame or tibble binding the rows of all data.Frames or tibbles included in sim
}
\description{
Bind simulated data of different situations
}
\examples{
\dontrun{
# Importing an example with three situations with observation:
workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situations= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situations)

bind_rows_sim(sim)
}
}
\seealso{
split_df2sim
}
