% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_situations.R
\name{cat_situations}
\alias{cat_situations}
\title{Format simulation data and observation data in order to consider all
situations}
\usage{
cat_situations(
  list_sim = NULL,
  obs = NULL,
  obs_sd = NULL,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{list_sim}{A list (each element= version) of a list
(each element= situation) of simulations \code{data.frame}s}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s
(named by situation)}

\item{obs_sd}{A list (each element= situation) of \code{data.frame}s for the
standard deviation of the observations (named by situation)}

\item{force}{Continue if the plot is not possible ? E.g. no observations for
scatter plots. If \code{TRUE}, return \code{NULL}, else return an error (default).}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
A list of three : a list (each element=version) of a list of a single
simulations \code{data.frame} named "all_situations", a list of a single
observations \code{data.frame} named "all_situations", and the same for obs_sd.
}
\description{
Format simulation data and observation data in a list of a
single data frame named "all_situations"
}
\keyword{internal}
