% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_hardWrapper}
\alias{ensure_hardWrapper}
\title{Ensure the existence of a list of variables in a \code{cropr_input} object}
\usage{
ensure_hardWrapper(data_object, parameters, type, supp_args = NULL)
}
\arguments{
\item{data_object}{A \code{cropr_input} object}

\item{parameters}{A list of characters}

\item{type}{A character}

\item{supp_args}{A list of varaibles:threshold_Tmax et threshold_Tmin}
}
\value{
A \code{cropr_input} object containing all variables in \code{parameters}.
}
\description{
Ensure the existence of a list of variables in a \code{cropr_input} object
}
\details{
This function will throw an error if any of the parameters can not be calculated with the given \code{data_object}.

The \code{type} argument is only used to make the error message more specific.

The name of the variable passed in \code{data_object} is captured and used for an error message.
}
\keyword{internal}
