% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_softWrapper}
\alias{ensure_softWrapper}
\title{Ensure the existence of a list of variables in a \code{cropr_input} object without provocing an error}
\usage{
ensure_softWrapper(data_object, parameters)
}
\arguments{
\item{data_object}{A \code{cropr_input} object}

\item{parameters}{A list of variables that we are interested in, for example: thickness, soil_max_wtr_cap, organic_N_conc}
}
\value{
A \code{cropr_input} object containing all variables in \code{parameters} if calculation was successfull.
\code{NULL} if any of the demanded parameters could not be calculated.
}
\description{
Ensure the existence of a list of variables in a \code{cropr_input} object without provocing an error
}
\keyword{internal}
