% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plot.R
\name{extract_plot}
\alias{extract_plot}
\title{Extract plot(s) from ggplot}
\usage{
extract_plot(plot, var = NULL, verbose = TRUE)
}
\arguments{
\item{plot}{The output of plot_situations}

\item{var}{A list of variables to extract}

\item{verbose}{Boolean. Print information during execution.}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}
}
\value{
A (printed) list of ggplot objects, each element being a plot for a situation
}
\description{
Extract a plot corresponding to one or several variables from a ggplot.
}
