% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_data.R
\name{get_argValues}
\alias{get_argValues}
\title{Get current values of function arguments}
\usage{
get_argValues()
}
\value{
A list containing the values of the parent function's arguments at that \code{get_argValues} is called.
}
\description{
Get current values of function arguments
}
