% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_plot_situations_arguments.R
\name{parse_plot_situations_args}
\alias{parse_plot_situations_args}
\title{Parse plot situations arguments}
\usage{
parse_plot_situations_args(
  dot_args,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = NULL,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  successive = NULL,
  shape_sit = c("none", "txt", "symbol", "group"),
  situation_group = NULL,
  reference_var = NULL,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obs}{A list (each element= situation) of observations \code{data.frame}s
(named by situation)}

\item{obs_sd}{A list (each element= situation) of standard deviations of
observations \code{data.frame}s (named by situation)}

\item{type}{The type of plot requested, either "dynamic" (date in X, variable
in Y) or scatter (simulated VS observed)}

\item{select_dyn}{Which data to plot when \code{type= "dynamic"}? See details.}

\item{select_scat}{Which data to plot when \code{type= "scatter"}? See details.}

\item{var}{A vector of variables that should be displayed on the graph.}

\item{title}{A vector of plot titles, named by situation. Use the situation
name if \code{NULL}, recycled if length one.}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, plot all
situations on the same graph.}

\item{overlap}{A list of lists containing the variables to represent on the
same graph when \code{type = "dynamic"}.}

\item{successive}{A list of lists containing the situations to be represented
as a contiguous sequence when \code{type = "dynamic"}
(implies that the situations are correctly ordered).}

\item{shape_sit}{Shape to differentiate between situations when
\code{all_situations= TRUE}. See details.}

\item{situation_group}{A list of lists of situations to gather when
\code{shape_sit= "group"}.}

\item{reference_var}{Variable selected on x-axis when type is scaétter and
select_scat is res. It is possible to select
between observation and simulation of the reference variable.
(examples : reference_var = "lai_n_obs", reference_var = "mafruit_sim")}

\item{force}{Continue if the plot is not possible ? E.g. no observations for
scatter plots. If \code{TRUE}, return \code{NULL}, else return an error (default).}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
A list of parsed arguments, eventually corrected according to the
compatibility between options.
}
\description{
This function parses the arguments for \code{plot_situations}.
}
\keyword{internal}
