% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_plotting.R
\name{plot.stics_simulation}
\alias{plot.stics_simulation}
\alias{autoplot.stics_simulation}
\title{Plot situations by group of simulation}
\usage{
\method{plot}{stics_simulation}(
  ...,
  obs = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = NULL,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  rotation = NULL,
  force = TRUE,
  verbose = TRUE
)

autoplot.stics_simulation(
  ...,
  obs = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = NULL,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  rotation = NULL,
  force = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{type}{The type of plot requested, either "dynamic" (date in X, variable in Y) or scatter (simulated VS observed)}

\item{select_dyn}{Which data to plot when \code{type= "dynamic"}? See details.}

\item{select_scat}{Which data to plot when \code{type= "scatter"}? See details.}

\item{var}{A vector of variables that should be displayed on the graph.}

\item{title}{A vector of plot titles, named by situation. Use the situation name if \code{NULL}, recycled if length one.}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, plot all situations on the same graph.}

\item{overlap}{A list of lists containing the variables to represent on the same graph
when \code{type = "dynamic"}.}

\item{rotation}{A list of lists containing the situations to be represented as a contiguous sequence
when \code{type = "dynamic"} (implies that the situations are correctly ordered).}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}
}
\value{
A (printed) list of ggplot objects, each element being a plot for a situation
}
\description{
simulation outputs for one or several situations with or without observations, eventually grouped
by a model version (or any group actually)
}
\details{
The \code{select_dyn} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with observations, and simulations outputs when there are some
\item "all": all variables with any observations or simulations outputs
}

The \code{select_scat} argument can be:
\itemize{
\item "sim" (the default): plots observations in X and simulations in Y.
\item "res": plots observations in X and residuals (observations-simulations) in Y.
}
}
\note{
The plots titles are given by their situation name.
}
\examples{
\dontrun{
# Importing an example with three situations with observation:
workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situations= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situations)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situations)

plot(sim,obs=obs)
}
}
