% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plotting.R
\name{plot_generic_input}
\alias{plot_generic_input}
\title{Generate a plot based on any input characteristics}
\usage{
plot_generic_input(
  type,
  soil,
  weather,
  symbol = c("auto", "Year", "Site"),
  supp_args = NULL,
  situation,
  histogram = NULL,
  interactive = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of plot to be generated. Must be one of the plot
types supported by the package. Possibilities include all the \code{type} arguments of
other plot functions.}

\item{soil}{A soil data object.}

\item{weather}{A weather data}

\item{supp_args}{A list of supplementary arguments depending on the plot function called}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{...}{Arguments to pass on to the specific plot function, see details.}

\item{symbol=c("auto", "Year", "Site")}{The \code{symbol} argument can be:
\itemize{
\item "auto" (the default): one year is symboled by one shape and one station is symboled by one color.
\item "Year": one year is symboled by one shape and  all the stations is symboled by one same color .
\item "Site": one station is symboled by one color and  all the years is symboled by one same shape ..
\item "group": One shape for each group of years described by the user
}}
}
\value{
The plot of type \code{type}.
}
\description{
Generates a plot of type \code{type} reflecting input charactersitics of
any of the data objects (soil, weather, ...).
}
\details{
Use the \code{\link{set_soil}}, \code{{set_weather}} and \code{{set_situation}}
functions to create respective data objects from user-given data. ToDo: Add detail for ...
}
\examples{
\dontrun{
ToDo
}
}
