% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_plotting.R
\name{plot_generic_input}
\alias{plot_generic_input}
\title{Generate a plot based on any input characteristics}
\usage{
plot_generic_input(
  type,
  soil,
  weather,
  situation,
  histogram = NULL,
  interactive = NULL,
  verbose,
  ...
)
}
\arguments{
\item{type}{The type of plot to be generated. Must be one of the plot
types supported by the package. Possibilities include all the \code{type} arguments of
other plot functions.}

\item{soil}{A soil data object.}

\item{weather}{A weather data object.}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{...}{Arguments to pass on to the specific plot function, see details.}
}
\value{
The plot of type \code{type}.
}
\description{
Generates a plot of type \code{type} reflecting input charactersitics of
any of the data objects (soil, weather, ...).
}
\details{
Use the \code{\link{data_soil}}, \code{{data_weather}} and \code{{data_situation}}
functions to create respective data objects from user-given data. ToDo: Add detail for ...
}
\examples{
\dontrun{
ToDo
}
}
