% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ploting.R
\name{plot_generic_situation}
\alias{plot_generic_situation}
\title{Generic ploting of a situation}
\usage{
plot_generic_situation(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = var,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  rotation = NULL,
  force = TRUE,
  verbose = TRUE,
  formater
)
}
\arguments{
\item{sim}{A simulation data.frame}

\item{obs}{An observation data.frame (variable names must match)}

\item{select_dyn}{Which data to plot when \code{type= "dynamic"}? See details.}

\item{select_scat}{Which data to plot when \code{type= "scatter"}? See details.}

\item{var}{A vector of variables that should be displayed on the graph.}

\item{title}{The plot title}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, plot all situations on the same graph.}

\item{overlap}{A list of lists containing the variables to represent on the same graph
when \code{type = "dynamic"}.}

\item{rotation}{A list of lists containing the situations to be represented as a contiguous sequence
when \code{type = "dynamic"} (implies that the situations are correctly ordered).}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here.}
}
\value{
A ggplot object
}
\description{
Plots outputs of a model (and observations) for one situation. This function is used as a generic
ploting function for any models. To use it with your own model, please provide a wrapper function around your model
to format the outputs used by this function (see \code{\link[=format_stics]{format_stics()}} for a template), and then provide your custom function as
an argument to this one.
}
\details{
The \code{select_dyn} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common (used when \code{type= "scatter"} )
\item "obs": all variables with observations, and simulations outputs when there are some
\item "all": all variables with any observations or simulations outputs
}

The \code{select_scat} argument can be:
\itemize{
\item "sim" (the default): plots observations in X and simulations in Y.
\item "res": plots observations in X and residuals (observations-simulations) in Y.
}
}
\keyword{internal}
