% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ploting.R
\name{plot_generic_situation}
\alias{plot_generic_situation}
\title{Generic ploting of a situation}
\usage{
plot_generic_situation(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "all"),
  title = NULL,
  formater
)
}
\arguments{
\item{sim}{A simulation data.frame}

\item{obs}{An observation data.frame (variable names must match)}

\item{type}{The type of plot required, either "dynamic" or "scatter"}

\item{plot}{The priority to either simulation or observation points if missing values (see details)}

\item{title}{The plot title}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here.}
}
\value{
A ggplot object
}
\description{
Plots outputs of a model (and observations) for one situation. This function is used as a generic
ploting function for any models. To use it with your own model, please provide a wrapper function around your model
to format the outputs used by this function (see \code{\link[=format_stics]{format_stics()}} for a template), and then provide your custom function as
an argument to this one.
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim": all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common
\item "obs": all variables with obervations, and simulations outputs when there are some
\item "all": all variables with any obervations or simulations outputs
}
}
\keyword{internal}
