% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ploting.R
\name{plot_situations}
\alias{plot_situations}
\title{Generic plotting function for all models}
\usage{
plot_situations(
  ...,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "res", "all"),
  title = NULL,
  all_situations = TRUE,
  force = TRUE,
  verbose = TRUE,
  formater
)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{type}{The type of plot requested, either "dynamic" (date in X, variable in Y) or scatter (simulated VS observed)}

\item{plot}{Which data to plot in priority when \code{type= "dynamic"}? See details.}

\item{title}{A vector of plot titles, named by situation. Use the situation name if \code{NULL}, recycled if length one.}

\item{all_situations}{Boolean. If \code{TRUE}, plot all situations on the same graph.}

\item{force}{Continue if the plot is not possible ? E.g. no observations for scatter plots. If \code{TRUE}, return \code{NULL}, else return an error.}

\item{verbose}{Boolean. Print information during execution.}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here (see \code{\link[=plot_generic_situation]{plot_generic_situation()}} and \code{\link[=format_stics]{format_stics()}} for more information).}
}
\value{
A (printed) list of ggplot objects, each element being a plot for a situation
}
\description{
Plots simulation outputs for one or several situations with or without observations, eventually grouped
by a model version (or any group actually).
}
\details{
The \code{plot} argument can be:
\itemize{
\item "sim" (the default): all variables with simulations outputs, and observations when there are some
\item "common": variables with simulations outputs and observations in common (used when \code{type= "scatter"} )
\item "obs": all variables with observations, and simulations outputs when there are some
\item "res": variables with simulations outputs and observations in common to plot residuals (used when \code{type= "scatter"} )
\item "all": all variables with any observations or simulations outputs
}
}
\note{
The plots titles are given by their situation name.
}
\keyword{internal}
