% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting.R
\name{plot_soil}
\alias{plot_soil}
\title{Generate a plot based on soil characteristics}
\usage{
plot_soil(
  soil,
  type = "all",
  weather = NULL,
  situation = NULL,
  histogram = NULL,
  interactive = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{soil}{A soil data object.}

\item{type}{The type of plot to be generated. Possibilities include
\itemize{
\item \code{"thickness.mswc"} -- Soil thickness agains soil maximum water content, colour shows soil organic nitrogen concentration..
}}

\item{weather}{A weather data object.}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{...}{Arguments to pass on to the specific plot function.}
}
\value{
The plot of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a plot of type \code{type} reflecting soil charactersitics of
the soil data object \code{soil}, possibly complemented by information from other
data objects (eg. weather, situation).
}
\details{
Use the \code{\link{set_soil}}, \code{{set_weather}}
functions to create respective data objects from user-given data.
See vignette 'List of supplied plot types' for all possible plot types.
}
\examples{
# load example data
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
soil_data <- readRDS(file.path(workspace, "soil_data_wide.rds"))
soil <- set_soil(
    soil_data,
    id = "name",
    layer_thickness = list("epc", "cm"),
    layer_water_field_cap = list("HCCF", "g/g"),
    layer_water_wilting_pt = list("HMINF", "g/g"),
    layer_bulk_density = list("DAF", "g/cm^3"),
    organic_N_conc = list("norg", "g/g")
)

# create plot
p <- plot_soil(soil, type = "thickness.mswc")

}
