% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting.R
\name{plot_soil}
\alias{plot_soil}
\title{Generate a plot based on soil characteristics}
\usage{
plot_soil(
  soil,
  type = "all",
  weather = NULL,
  situation = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{soil}{A soil data object.}

\item{type}{The type of plot to be generated. Possibilities include
\itemize{
\item \code{"\link{thickness.mswc.norg}"} -- Thickness, maximum  water content and organic nitrogen per soil.
\item \code{"type2"} -- Another plot.
}}

\item{weather}{A weather data object.}

\item{situation}{A situation data object.}

\item{...}{Arguments to pass on to the specific plot function, see details.}
}
\value{
The plot of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a plot of type \code{type} reflecting soil charactersitics of
the soil data object \code{soil}, possibly complemented by information from other
data objects (eg. weather, situation).
}
\details{
Use the \code{\link{data_soil}}, \code{{data_weather}} and \code{{data_situation}}
functions to create respective data objects from user-given data. ToDo: Add detail for ...
}
\examples{
\dontrun{
# ToDo: add data that makes this example work
workspace= "path_to_workspace"

soil_data <- SticsRFiles::get_param_xml(file.path(workspace, "sols.xml"))[[1]]
soil_data$name <- SticsRFiles::get_soils_list(file.path(workspace, "sols.xml"))

soil_data$MSWC <- soil_data$epc*(soil_data$HCCF-soil_data$HMINF)
soil_data$MSWC <- colMeans(matrix(soil_data$MSWC, nrow=5))
soil_data$thickness <- colSums(matrix(soil_data$epc, nrow=5))
soil <- data_soil(data = soil_data, thickness=thickness, mswc=MSWC, norg=norg, name=name)

plot_soil(soil, type="thickness.mswc")
}

}
