% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting - Copie.R, R/cropr_plotting.R,
%   R/plots_weather.R
\name{plot_weather}
\alias{plot_weather}
\alias{plot__limiting.temperatures}
\title{Generate a plot based on weather characteristics}
\usage{
plot_weather(
  weather,
  type = "all",
  threshold_Tmin = 0,
  threshold_Tmax = 35,
  situation = NULL,
  histogram = NULL,
  interactive = NULL,
  verbose = FALSE,
  ...
)

plot_weather(
  weather,
  type = "all",
  threshold_Tmin = 0,
  threshold_Tmax = 35,
  situation = NULL,
  histogram = NULL,
  interactive = NULL,
  verbose = FALSE,
  ...
)

plot__limiting.temperatures(
  weather,
  histogram,
  interactive,
  threshold_Tmin,
  threshold_Tmax,
  ...
)
}
\arguments{
\item{weather}{A weather data object.}

\item{type}{The type of plot to be generated. Possibilities include
\itemize{
\item \code{"limiting.temperatures"} -- Number of hot days against number of cold days, colour and form show station site and year.
}}

\item{threshold_Tmin}{The threshold temperature that defines cold days, =0 by default}

\item{threshold_Tmax}{The threshold temperature that defines hot days. =35 by default}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{...}{Arguments to pass on to the specific plot function.}
}
\value{
The plot of type \code{type} with data from on the weather data object \code{weather}.

The plot of type \code{type} with data from on the weather data object \code{weather}.
}
\description{
Generates a plot of type \code{type} reflecting weather characteristics of
the weather data object \code{weather}, possibly complemented by information from other
data objects (eg. situation).

Generates a plot of type \code{type} reflecting weather characteristics of
the weather data object \code{weather}, possibly complemented by information from other
data objects (eg. situation).
}
\details{
Use the \code{\link{set_weather}}, \code{{set_weather}}.
functions to create respective data objects from user-given data.
See vignette 'List of supplied plot types' for all possible plot types.

Use the \code{\link{set_weather}}, \code{{set_weather}}.
functions to create respective data objects from user-given data.
See vignette 'List of supplied plot types' for all possible plot types.
}
\examples{
# load example data
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
weather_data <- readRDS(file.path(workspace, "weather_data_list.rds"))

weather <- set_weather(
    weather_data,
    id = "station",
    temp_day_max = list("ttmax", "celsius"),
    temp_day_min = list("ttmin", "celsius"),
    year = "year"
)

# create plot
p <- plot_weather(weather, type = "limiting.temperatures")
# load example data
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
weather_data <- readRDS(file.path(workspace, "weather_data_list.rds"))

weather <- set_weather(
    weather_data,
    id = "station",
    temp_day_max = list("ttmax", "celsius"),
    temp_day_min = list("ttmin", "celsius"),
    year = "year"
)

# create plot
p <- plot_weather(weather, type = "limiting.temperatures")
}
\keyword{internal}
