% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting.R, R/plots_weather.R
\name{plot_weather}
\alias{plot_weather}
\alias{plot__limiting.temperatures}
\alias{plot__limiting.rainfall_days}
\alias{plot__temperature.rainfall}
\alias{plot__temperature}
\alias{plot__radiation}
\alias{plot__cumulated_PET.rain}
\title{Generate a plot based on weather characteristics}
\usage{
plot_weather(
  weather,
  type = "all",
  symbol = c("auto", "Year", "Site"),
  threshold_Tmin = 0,
  threshold_Tmax = 35,
  threshold_RainMin = -5,
  threshold_RainMax = 20,
  situation = NULL,
  histogram = NULL,
  interactive = FALSE,
  verbose = FALSE,
  cumulate = FALSE,
  ...
)

plot__limiting.temperatures(
  weather,
  histogram,
  symbol = c("auto", "Year", "Site"),
  interactive,
  threshold_Tmin,
  threshold_Tmax,
  ...
)

plot__limiting.rainfall_days(
  weather,
  histogram = NULL,
  symbol = c("auto", "Year", "Site"),
  interactive,
  threshold_RainMin,
  threshold_RainMax,
  ...
)

plot__temperature.rainfall(
  weather,
  histogram = NULL,
  symbol = c("auto", "Year", "Site"),
  interactive,
  ...
)

plot__temperature(weather, histogram = NULL, ...)

plot__radiation(weather, histogram = NULL, cumulate = FALSE, ...)

plot__cumulated_PET.rain(
  weather,
  histogram = NULL,
  symbol = c("auto", "Year", "Site"),
  interactive,
  ...
)
}
\arguments{
\item{weather}{A weather data object.}

\item{type}{The type of plot to be generated. Possibilities include
\itemize{
\item \code{"limiting.temperatures"} -- Number of hot days against number of cold days, colour and form show station site and year.
\item \code{"temperature.rainfall"} -- annual average temperature against cumulated rainfall  , colour and form show station site and year.
\item \code{"limiting.rainfall_days"} -- Number of dry days against number of high rainfall days, colour and form show station site and year.
\item \code{"radiation"} --  global radiation of each situation against julian days , site and year.
\item \code{"temperature"} -- temperature (min,max,mean)against julian days, site and year.
\item \code{"cumulated_PET.rain"} -- cumulated_PET" vs "cumulated_rain", colour and form show station site and year.
}}

\item{threshold_Tmin}{The threshold temperature that defines cold days, =0 by default}

\item{threshold_Tmax}{The threshold temperature that defines hot days. =35 by default}

\item{threshold_RainMin}{The threshold rainfall that defines dry days, =-5 by default}

\item{threshold_RainMax}{The threshold rainfall that defines high rainfall days, =20 by default}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{cumulate}{if cumulate= TRUE, we cumulate the radiation,else FALSE we don't cumulate it, =FALSE by default}

\item{...}{Arguments to pass on to the specific plot function.}

\item{symbol=c("auto", "Year", "Site")}{The \code{symbol} argument can be:
\itemize{
\item "auto" (the default): one year is symboled by one shape and one station is symboled by one color.
\item "Year": one year is symboled by one shape and  all the stations is symboled by one same color .
\item "Site": one station is symboled by one color and  all the years is symboled by one same shape .
\item "list(…)": A list of lists containing the years or stations to be clustered ,and one shape for each group of years(stations) described by the user.
}}
}
\value{
The plot of type \code{type} with data from on the weather data object \code{weather}.
}
\description{
Generates a plot of type \code{type} reflecting weather characteristics of
the weather data object \code{weather}, possibly complemented by information from other
data objects (eg. situation).
}
\details{
Use the \code{\link{set_weather}}, \code{{set_weather}}.
functions to create respective data objects from user-given data.
See vignette 'List of supplied plot types' for all possible plot types.
}
\examples{
# load example data
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
weather_data <- readRDS(file.path(workspace, "weather_data_list.rds"))

weather <- set_weather(
    weather_data,
    id = "station",
    temp_day_max = list("ttmax", "celsius"),
    temp_day_min = list("ttmin", "celsius"),
    year = "year",
    rainfall_day = list("ttrr", "mm"),
    radiation_day = list("ttrg", "MJ.m-2"),
    etp_day = list("ttetp", "mm")
)

# create plot
p <- plot_weather(weather, type = "limiting.temperatures")
p <- plot_weather(weather, type = "temperature.rainfall")
}
\keyword{internal}
