% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_plotting.R
\name{plot_weather}
\alias{plot_weather}
\title{Generate a plot based on soil characteristics}
\usage{
plot_weather(
  weather,
  type = "all",
  soil = NULL,
  situation = NULL,
  histogram = NULL,
  interactive = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{weather}{A weather data object.}

\item{type}{The type of plot to be generated. Possibilities include
\itemize{
\item \code{"\link{limiting.temeratures}"}
}}

\item{soil}{A weather data object.}

\item{situation}{A situation data object.}

\item{histogram}{Draw graph in histogram-like form?}

\item{interactive}{Transform output to an interactive \code{plotly} plot?}

\item{verbose}{Print details on the console while executing the function?}

\item{...}{Arguments to pass on to the specific plot function, see details.}
}
\value{
The plot of type \code{type} with data from on the soil data object \code{soil}.
}
\description{
Generates a plot of type \code{type} reflecting soil charactersitics of
the soil data object \code{soil}, possibly complemented by information from other
data objects (eg. weather, situation).
}
\details{
Use the \code{\link{set_soil}}, \code{{set_weather}}.
functions to create respective data objects from user-given data. ToDo: Add detail for ...
}
