% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_soil.R
\name{set_soil}
\alias{set_soil}
\title{Collect data on soil variables for further use in plot-functions}
\usage{
set_soil(
  data,
  id,
  thickness = NULL,
  soil_max_wtr_cap = NULL,
  organic_N_conc = NULL,
  layer_thickness = NULL,
  layer_water_field_cap = NULL,
  layer_water_wilting_pt = NULL,
  layer_bulk_density = NULL,
  layer_max_wtr_cap = NULL,
  data_format = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame containing soil data}

\item{id}{Soil identification}

\item{thickness}{Soil thickness}

\item{soil_max_wtr_cap}{Soil maximal water capacity}

\item{organic_N_conc}{Soil organic nitrogen content}

\item{layer_thickness}{Soil thickness by layer}

\item{layer_water_field_cap}{Soil water field capacity by layer}

\item{layer_water_wilting_pt}{Soil water wilting point by layer}

\item{layer_bulk_density}{Soil bulk density of fine earth fraction by layer}

\item{layer_max_wtr_cap}{Soil maximal water capacity by layer}

\item{data_format}{Specify format of data argument, see \code{Details} for usage}

\item{verbose}{Provide extra information about the function's inner procedures?}
}
\value{
A list of class \code{cropr_input} containing all necessary information for plotting.
}
\description{
Transforms user-provided soil data into an object that can be read by other CroPlotR
functions, most notably by the plot functions. The data on all variables has to be given through
a single data frame. The names of every variable and its unit have to be specified.
}
\details{
The column name in \code{data} and the unit of every variable have to be given as a list of two characters.

Possible values of the \code{data_format} argument are "tibble", "wide" and "long", see vignette (link?) for details.
If no data format is specified, an automatic detection is attempted. In the case of the \emph{long} data frame format, the
\code{data_format} argument can be a list where the first element determines the format and the second element is a named list.
The elements in this named list must be called 'id', 'variable', 'layer' and 'value' and they must be either integers or characters.
They specify by index or by names in which column of the long data frame the identifications, variables, layers and values
are to be found. If the element 'id' is not specified, the regular 'id' argument of \code{set_soil} is used.
}
\examples{
# load example data in different formats
workspace <- system.file(file.path("extdata", "example_input"), package = "CroPlotR")
soil_data_wide <- readRDS(file.path(workspace, "soil_data_wide.rds"))
soil_data_long <- readRDS(file.path(workspace, "soil_data_long.rds"))
soil_data_tibble <- readRDS(file.path(workspace, "soil_data_tibble.rds"))

# usage with automatic format detection
soil <- set_soil(
  soil_data_wide,
  id = "name",
  organic_N_conc = list("norg", "g/g"),
  layer_thickness = list("epc", "cm")
)

# usage with specific format
soil <- set_soil(
  soil_data_wide,
  id = "name",
  organic_N_conc = list("norg", "g/g"),
  layer_thickness = list("epc", "cm"),
  data_format = "wide"
)

# usage in long format while specifying columns
soil <- set_soil(
  soil_data_long,
  id = "name",
  organic_N_conc = list("norg", "g/g"),
  layer_thickness = list("epc", "cm"),
  data_format = list("long", list(id = "name", variable = "variable", layer = 3, value = "value"))
)

}
