% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plotting_dynamic.R
\name{specific_dynamic_plots}
\alias{specific_dynamic_plots}
\alias{plot_dynamic}
\title{Specific functions to generate dynamic plots}
\usage{
plot_dynamic(df_data, sit, title = NULL)
}
\arguments{
\item{df_data}{A named list of data frame including the data to plot (one df
per situation)}

\item{sit}{The name of the situation to plot}

\item{title}{The plot title (optional, NULL by default)}
}
\value{
A ggplot object

A list of ggplot objects
}
\description{
Generate dynamic plots for the differents
cases handled in CroPlotR
(plant mixture, plot several simulation results on same graph, ...)
as specitifed by the different arguments.
}
\details{
List of the different specific functions:
\itemize{
\item \code{plot_dynamic_mixture}: Generate a dynamic plot for the case of
mixture of crops, single simulation version and each variable
in different plot
\item \code{plot_dynamic_mixture_overlap}: Generate a dynamic plot for the
case of mixture of crops, single simulation version and several variables
in same plot (overlap)
\item \code{plot_dynamic_versions}: Generate a dynamic plot for the
case of sole crops, all simulation versions in same plot and each
variable in a plot
\item \code{plot_dynamic_overlap}: Generate a dynamic plot for the
case of sole crops, single simulation version and several variables
in same plot (overlap)
\item \code{plot_dynamic_mixture_versions_overlap}: not implemented (seems not
useful because too complex to analyze)
\item \code{plot_dynamic_versions_overlap}: Generate a dynamic plot for the
case of sole crops, all simulation versions and several variables
in same plot (overlap)
\item \code{plot_dynamic_mixture_versions}: Generate a dynamic plot for the
case of mixture of crops, all simulation versions in same plot and
each variable in a plot
}
}
\keyword{internal}
