% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plotting_scatter.R
\name{specific_scatter_plots}
\alias{specific_scatter_plots}
\alias{plot_scat_mixture_allsit}
\title{Specific functions to generate scatter plots}
\usage{
plot_scat_mixture_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)
}
\arguments{
\item{df_data}{A named list of data frame including the data to plot (one df
per situation, or only one df if sit==all_situations)}

\item{sit}{The name of the situation to plot (or all_situations)}

\item{is_obs_sd}{TRUE if error standard deviation of observations is provided}
}
\value{
A ggplot object

A list of ggplot objects
}
\description{
Generate scatter plots for the differents cases handled in CroPlotR
(plant mixture, plot of residuals, plot several simulation results on same graph, ...)
as specitifed by the different arguments.
}
\details{
List of the different specific functions:
\itemize{
\item \code{plot_scat_mixture_allsit}: Generate a scatter plot for the case of
mixture of crops, single simulation version and all_situations in same plot
}
}
\keyword{internal}
