% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_plotting_scatter.R
\name{specific_scatter_plots}
\alias{specific_scatter_plots}
\alias{compute_axis_bounds}
\alias{make_axis_square}
\alias{add_obs_error_bars}
\alias{plot_scat_mixture_allsit}
\alias{plot_scat_allsit}
\title{Specific functions to generate scatter plots}
\usage{
compute_axis_bounds(df_data, reference_var, y_var_type, is_obs_sd)

make_axis_square(df_data, reference_var, y_var_type, is_obs_sd, p)

add_obs_error_bars(p, colour_factor)

plot_scat_mixture_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)

plot_scat_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL,
  several_sit = FALSE,
  one_version = FALSE,
  mixture = FALSE
)
}
\arguments{
\item{df_data}{A named list of data frame including the data to plot (one df
per situation, or only one df if sit==all_situations)}

\item{y_var_type}{type of variable to plot ("Simulated" or "Residuals")}

\item{is_obs_sd}{TRUE if error standard deviation of observations is provided}

\item{p}{A ggplot to modify`}

\item{colour_factor}{The factor to use for colouring the error bars}

\item{sit}{The name of the situation to plot (or all_situations)}

\item{several_sit}{TRUE if the plot is for several situations}

\item{one_version}{TRUE if the plot is for one version}

\item{mixture}{TRUE if the plot is for a mixture of crops}
}
\value{
A ggplot object

A list of ggplot objects

List of x and y axis bounds (xaxis_min, xaxis_max, yaxis_min,
yaxis_max)

The modified ggplot

The modified ggplot
}
\description{
Generate scatter plots for the different cases handled in
CroPlotR (plant mixture, plot of residuals, plot several simulation results
on same graph, ...) as specified by the different arguments.

Compute axis bounds (+/-0.05 added to the min/max of the data).

Make axis square

Add error bars on observed values in given scatterplot
}
\details{
List of the different specific functions:
\itemize{
\item \code{plot_scat_mixture_allsit}: Generate a scatter plot for the case of
mixture of crops, single simulation version and all_situations in same plot
\item \code{plot_scat_allsit}: Generate a scatter plot for the case of
sole crops, single simulation version and all_situations in same plot
}
}
\keyword{internal}
