% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stats.R
\name{statistics_situations}
\alias{statistics_situations}
\title{Generic statistics for all situations}
\usage{
statistics_situations(..., obs = NULL, formater)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here (see \code{\link[=statistics]{statistics()}} and \code{\link[=format_stics]{format_stics()}} for more information).}
}
\value{
A list of statistics \code{data.frame}s named by situation
}
\description{
Compute statistics for simulations of one or several situations with its observations, eventually grouped
by a model version (or any group actually).
}
\keyword{internal}
