% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_stats.R
\name{statistics_situations}
\alias{statistics_situations}
\title{Generic statistics for all situations}
\usage{
statistics_situations(
  ...,
  obs = NULL,
  stat = "all",
  all_situations = TRUE,
  all_plants = TRUE,
  verbose = TRUE,
  formater
)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{stat}{A character vector of required statistics, "all" for all, or any of \code{\link[=predictor_assessment]{predictor_assessment()}}.}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, computes statistics for all situations.}

\item{all_plants}{Boolean (default = TRUE). If \code{TRUE}, computes statistics for all plants (when applicable).}

\item{verbose}{Boolean. Print information during execution.}

\item{formater}{The function used to format the models outputs and observations in a standard way. You can design your own function
that format one situation and provide it here (see \code{\link[=statistics]{statistics()}} and \code{\link[=format_cropr]{format_cropr()}} for more information).}
}
\value{
A \code{\link[tibble:as_tibble]{tibble::as_tibble()}} with statistics grouped by group (i.e. model version) and situation
}
\description{
Compute statistics for simulations of one or several situations with its observations, eventually grouped
by a model version (or any group actually).
}
\seealso{
All the functions used to compute the statistics: \code{\link[=predictor_assessment]{predictor_assessment()}}.
}
\keyword{internal}
