% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropr_stats.R
\name{summary.cropr_simulation}
\alias{summary.cropr_simulation}
\title{Summary statistics of simulations}
\usage{
\method{summary}{cropr_simulation}(
  ...,
  obs,
  stats = "all",
  all_situations = TRUE,
  verbose = TRUE,
  stat = lifecycle::deprecated()
)
}
\arguments{
\item{...}{Simulation outputs (each element= model version), each being a named list of \code{data.frame} for each situation.
See examples.}

\item{obs}{A list (each element= situation) of observations \code{data.frame}s (named by situation)}

\item{stats}{A character vector of required statistics, "all" for all, or any of \code{\link[=predictor_assessment]{predictor_assessment()}}.}

\item{all_situations}{Boolean (default = TRUE). If \code{TRUE}, computes statistics for all situations.}

\item{verbose}{Logical value for displaying information while running}

\item{stat}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{stat} is no
longer supported, use \code{stats} instead.#'}
}
\value{
A list of statistics \code{data.frame}s named by situation
}
\description{
Summary statistics for one or several situations with observations, eventually grouped
by a model version (or any group actually)
}
\examples{
\dontrun{
# Importing an example with three situations with observation:
workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situations= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))
sim= SticsRFiles::get_sim(workspace = workspace, usm = situations)
obs= SticsRFiles::get_obs(workspace =  workspace, usm = situations)

# All stats for the simulation:
summary(sim, obs= obs)

# All stats for two groups of simulations:
summary(sim1= sim, sim2= sim, obs=obs)

# Only R2 and nRMSE for one group:
summary(sim, obs=obs, stats= c("R2","nRMSE"))

}
}
\seealso{
All the functions used to compute the statistics: \code{\link[=predictor_assessment]{predictor_assessment()}}.
}
